+++ src/mma/merge/__init__.py
@@ -1,6 +1,15 @@
 """Multitask merge package (P5)."""
 
 from mma.merge.merge_multitask import assert_no_missing_tasks, merge_multitask
+from mma.merge.merge_to_final import merge_to_final
 from mma.merge.validate_ready import validate_ready
+from mma.merge.write_final import FINAL_MANIFEST_NAME, write_final_manifest
 
-__all__ = ["assert_no_missing_tasks", "merge_multitask", "validate_ready"]
+__all__ = [
+    "FINAL_MANIFEST_NAME",
+    "assert_no_missing_tasks",
+    "merge_multitask",
+    "merge_to_final",
+    "validate_ready",
+    "write_final_manifest",
+]